refactor: 优化流读取和 Mongobar UI 渲染逻辑，调整 cost 图表的标记样式，优化压力测试细节

+++ src/mongobar/mongobar_config.rs
@@ -10,5 +10,5 @@ pub(crate) struct MongobarConfig {
     #[educe(Default = 10)]
     pub thread_count: u32,
     #[educe(Default = 1000)]
-    pub loop_count: i32,
+    pub loop_count: usize,
 }
